// packages/@ros-cdk/services/fnf/lib/fnf.generated.ts
export interface RosFlowProps {

    /**
     * @Property definition: The definition of the created flow following the FDL syntax standard.
     */
    readonly definition: string;

    /**
     * @Property name: The name of the flow created. This name is unique under the account.
     */
    readonly name: string;

    /**
     * @Property description: Create a description of the flow.
     */
    readonly description?: string;

    /**
     * @Property requestId: The specified Request ID for this request. If not specified, our system will help you generate a random one.
     */
    readonly requestId?: string;

    /**
     * @Property roleArn: Optional parameter, the resource descriptor information required for the execution of the flow, used to perform the assume role during FnF execution.
     */
    readonly roleArn?: string;
}
// ...
export class RosFlow extends ros.RosResource {
    /**
     * The resource type name for this resource class.
     */
    public static readonly ROS_RESOURCE_TYPE_NAME = "ALIYUN::FNF::Flow";

    /**
     * A factory method that creates a new instance of this class from an object
     * containing the properties of this ROS resource.
     */

    /**
     * @Attribute CreatedTime: Flow creation time.
     */
    public readonly attrCreatedTime: any;

    /**
     * @Attribute Id: The unique ID of the flow.
     */
    public readonly attrId: any;

    /**
     * @Attribute LastModifiedTime: The most recently modified time of the flow.
     */
    public readonly attrLastModifiedTime: any;

    /**
     * @Attribute Name: The name of the flow created.
     */
    public readonly attrName: any;

    public enableResourcePropertyConstraint: boolean;


    /**
     * @Property definition: The definition of the created flow following the FDL syntax standard.
     */
    public definition: string;

    /**
     * @Property name: The name of the flow created. This name is unique under the account.
     */
    public name: string;

    /**
     * @Property description: Create a description of the flow.
     */
    public description: string | undefined;

    /**
     * @Property requestId: The specified Request ID for this request. If not specified, our system will help you generate a random one.
     */
    public requestId: string | undefined;

    /**
     * @Property roleArn: Optional parameter, the resource descriptor information required for the execution of the flow, used to perform the assume role during FnF execution.
     */
    public roleArn: string | undefined;

    /**
     * Create a new `ALIYUN::FNF::Flow`.
     *
     * @param scope - scope in which this resource is defined
     * @param id    - scoped id of the resource
     * @param props - resource properties
     */
    constructor(scope: ros.Construct, id: string, props: RosFlowProps, enableResourcePropertyConstraint: boolean) {
        super(scope, id, { type: RosFlow.ROS_RESOURCE_TYPE_NAME, properties: props });
        this.attrCreatedTime = ros.Token.asString(this.getAtt('CreatedTime'));
        this.attrId = ros.Token.asString(this.getAtt('Id'));
        this.attrLastModifiedTime = ros.Token.asString(this.getAtt('LastModifiedTime'));
        this.attrName = ros.Token.asString(this.getAtt('Name'));

        this.enableResourcePropertyConstraint = enableResourcePropertyConstraint;
        this.definition = props.definition;
        this.name = props.name;
        this.description = props.description;
        this.requestId = props.requestId;
        this.roleArn = props.roleArn;
    }


    protected get rosProperties(): { [key: string]: any }  {
        return {
            definition: this.definition,
            name: this.name,
            description: this.description,
            requestId: this.requestId,
            roleArn: this.roleArn,
        };
    }
    protected renderProperties(props: {[key: string]: any}): { [key: string]: any }  {
        return rosFlowPropsToRosTemplate(props, this.enableResourcePropertyConstraint);
    }
}